+++ examples/hackerNewsAnalyzer/hackerNewsAnalyzer.tsx
@@ -1,4 +1,4 @@
-import { OpenAIProvider, TextCompletion } from "@gensx/openai";
+import { ChatCompletion, OpenAIProvider } from "@gensx/openai";
 import { gsx } from "gensx";
 
 import { getTopStoryDetails, type HNStory } from "./hn.js";
@@ -27,7 +27,7 @@ Focus on the most surprising or counterintuitive point rather than trying to sum
     `.trim();
 
     return (
-      <TextCompletion
+      <ChatCompletion
         messages={[
           { role: "system", content: PROMPT },
           {
@@ -66,7 +66,7 @@ Maintain your voice while preserving the key insights and all links from the ana
   `.trim();
 
     return (
-      <TextCompletion
+      <ChatCompletion
         messages={[
           { role: "system", content: PROMPT },
           { role: "user", content: content },
@@ -123,7 +123,7 @@ Focus on substance rather than surface-level reactions. When referencing comment
     .join("\n\n");
 
   return (
-    <TextCompletion
+    <ChatCompletion
       messages={[
         { role: "system", content: PROMPT },
         {
@@ -177,7 +177,7 @@ ${story.comments
     `.trim();
 
     return (
-      <TextCompletion
+      <ChatCompletion
         messages={[
           { role: "system", content: PROMPT },
           { role: "user", content: context },
@@ -191,7 +191,7 @@ ${story.comments
           }
           return response;
         }}
-      </TextCompletion>
+      </ChatCompletion>
     );
   },
 );
@@ -245,7 +245,7 @@ ${commentAnalysis}
       .join("\n\n");
 
     return (
-      <TextCompletion
+      <ChatCompletion
         messages={[
           { role: "system", content: PROMPT },
           { role: "user", content: context },
