+++ examples/openai-v-next/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@gensx-examples/hacker-news-analyzer",
+  "name": "@gensx-examples/openai-v-next",
   "private": true,
   "version": "0.0.0",
   "type": "module",
